adds homescreen starts dev on diffscreen
