--- app/src/main/java/goodguygregory/app/musictrainer/KeySignature.kt
@@ -1,5 +1,0 @@
-package goodguygregory.app.musictrainer
-
-class KeySignature {
-
-}
